# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Web+ API
  description: |
    Web+ is a comprehensive platform for managing and interacting with large language models.
    This API provides endpoints for authentication, model management, chat completions,
    file handling, and pipeline automation.
  version: 0.2.0
  contact:
    name: Web+ Support
    email: support@webplus.ai
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8002/api
    description: Local development server
  - url: https://api.webplus.ai
    description: Production server
security:
  - bearerAuth: []
  - apiKeyAuth: []
...
paths:
...
  /files/upload:
    post:
      tags: [Files]
      summary: Upload a file
      operationId: uploadFile
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                conversation_id:
                  type: string
      responses:
        '201':
          description: File uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileUploadResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
  schemas:
...
    FileUploadResponse:
      type: object
      required: [id, filename, size]
      properties:
        id:
          type: string
          format: uuid
        filename:
          type: string
        size:
          type: integer
        content_type:
          type: string
        upload_date:
          type: string
          format: date-time
...